maint: make sha512_reverse() available for non-SIMD
It must be available for a SIMD or non-SIMD build since it can be
used inside an OpenCL format.
